Added null checks in DeleteShopifyWebhookAsync

+++ affiliate-proj/Application/Services/Shopify/Webhook/ShopifyWebhookRepository.cs
@@ -50,7 +50,13 @@ public class ShopifyWebhookRepository : IShopifyWebhookRepository
         var id = shop.Id;
         var store = await _dbContext.Stores.FirstOrDefaultAsync(s => s.ShopifyId == id);
         
+        if (store == null)
+            return false;
+        
         var webhook = await _dbContext.WebhookRegistrations.FirstOrDefaultAsync(r => r.StoreId == store.StoreId);
+
+        if (webhook != null)
+            return false;
         
         _dbContext.WebhookRegistrations.Remove(webhook);
         await _dbContext.SaveChangesAsync();
